2010-02-09 15:43 UTC+0100 Viktor Szakats (harbour.01 syenar.hu)   * src/pp/ppcore.c   * src/vm/strapi.c   * src/vm/debug.c   * src/vm/itemapi.c   * src/vm/cmdarg.c   * src/vm/set.c   * src/debug/dbgentry.c   * src/common/hbfsapi.c   * src/common/hbstr.c   * src/common/strwild.c   * src/nortl/nortl.c   * src/rtl/lennum.c   * src/rtl/strmatch.c   * src/rtl/gtstd/gtstd.c   * src/rtl/hbstrfmt.c   * src/rtl/transfrm.c   * src/rtl/gtcgi/gtcgi.c   * src/rtl/filesys.c   * src/rtl/console.c   * src/rtl/hbgtcore.c   * src/rtl/cdpapi.c   * src/rtl/gtclip.c   * src/rtl/mlcfunc.c   * src/rtl/itemseri.c   * src/rtl/gtpca/gtpca.c   * src/rtl/samples.c   * src/rtl/gete.c   * src/rdd/workarea.c   * src/rdd/hsx/hsx.c   * src/rdd/hbsix/sxsem.c   * src/rdd/hbsix/sxfname.c   * contrib/xhb/dbf2txt.c   * contrib/xhb/datesxhb.c   * contrib/xhb/dbgfxc.c   * contrib/hbmzip/hbmzip.c   * contrib/hbmemio/memio.c   * contrib/hbnetio/netiocli.c   * contrib/gtwvg/wvgwing.c   * contrib/hbpgsql/postgres.c   * contrib/hbclipsm/num.c   * contrib/rddads/ads1.c   * contrib/hbfimage/fi_wrp.c     + Added '( HB_SIZE )' casts to 'strlen()' calls where       it was converted to HB_SIZE. This pacifies long standing       win64 warnings.       Remaining warnings are (among others) useful to spot places where       non-HB_SIZE is used in Harbour code to store sizes.
  * contrib/hbpgsql/postgres.c
    ! Type cleanup.

+++ harbour/src/common/hbfsapi.c
@@ -251,7 +251,7 @@ char * hb_fsFNameMerge( char * pszFileName, PHB_FNAME pFileName )
       was none. */
    if( pszFileName[ 0 ] != '\0' && ( pszName || pFileName->szExtension ) )
    {
-      int iLen = strlen( pszFileName ) - 1;
+      int iLen = ( int ) strlen( pszFileName ) - 1;
 
       if( iLen < HB_PATH_MAX - 1 - 2 && pszFileName[ iLen ] != cDirSep &&
           strchr( HB_OS_PATH_DELIM_CHR_LIST, pszFileName[ iLen ] ) == NULL )
